+++ ruleminer/parser.py
@@ -7,8 +7,49 @@ _lbra, _rbra = map(pyparsing.Suppress, "[]")
 # _sep = pyparsing.Suppress(",")
 _sep = pyparsing.Literal(",")
 _number = pyparsing.Regex(r"[+-]?\d+(?:\.\d*)?(?:[eE][+-]?\d+)?")
-_function = pyparsing.one_of(
-    "min \
+_datefunction = pyparsing.one_of(
+    "day_name \
+    month_name \
+    days_in_month \
+    daysinmonth \
+    is_leap_year \
+    is_year_end \
+    dayofweek \
+    weekofyear \
+    weekday \
+    week \
+    is_month_end \
+    is_month_start \
+    is_year_start \
+    is_quarter_end \
+    is_quarter_start \
+    day \
+    month \
+    quarter \
+    year \
+    DAY_NAME \
+    MONTH_NAME \
+    DAYS_IN_MONTH \
+    DAYSINMONTH \
+    IS_LEAP_YEAR \
+    IS_YEAR_END \
+    DAYOFWEEK \
+    WEEKOFYEAR \
+    WEEKDAY \
+    WEEK \
+    IS_MONTH_END \
+    IS_MONTH_START \
+    IS_YEAR_START \
+    IS_QUARTER_END \
+    IS_QUARTER_START \
+    DAY \
+    MONTH \
+    QUARTER \
+    YEAR"
+)
+_function = (
+    pyparsing.one_of(
+        "min \
     max \
     abs \
     quantile \
@@ -28,6 +69,8 @@ _function = pyparsing.one_of(
     COUNT \
     SUMIF \
     COUNTIF"
+    )
+    | _datefunction
 )
 _for = pyparsing.one_of("for", "FOR")
 _in = pyparsing.one_of("in", "IN")
@@ -90,10 +133,7 @@ def function_expression():
     param_condition = param_element + _compa_op + param_element
 
     param_condition_list = pyparsing.Group(
-        _lbra
-        + pyparsing.Group(param_condition)
-        + (_sep + pyparsing.Group(param_condition))[...]
-        + _rbra
+        _lbra + param_condition + (_sep + param_condition)[...] + _rbra
     )
     param_condition_list_comprehension = pyparsing.Group(
         _lbra
@@ -102,9 +142,8 @@ def function_expression():
         + _list_comprehension_var
         + _in
         + _lbra
-        + pyparsing.Group(
-            pyparsing.Group(_column) + (_sep + pyparsing.Group(_column))[...]
-        )
+        + _column
+        + (_sep + _column)[...]
         + _rbra
         + _rbra
     )